Remove the quiz export logic from the controller
Isolate it to make it more testeable and understandable

Refactor for #1801
